Move DownloadManagerWrapper to DownloadUtility. Move top-level functions.
