+++ app/src/main/java/tech/ula/model/remote/GithubAppsFetcher.kt
@@ -2,6 +2,8 @@ package tech.ula.model.remote
 
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import tech.ula.model.entities.App
 import tech.ula.utils.* // ktlint-disable no-wildcard-imports
 import java.io.BufferedReader
@@ -10,23 +12,11 @@ import java.io.IOException
 import java.io.InputStreamReader
 import java.net.URL
 
-interface RemoteAppsSource {
-    fun getHostname(): String
-
-    suspend fun fetchAppsList(): List<App>
-
-    suspend fun fetchAppIcon(app: App)
-
-    suspend fun fetchAppDescription(app: App)
-
-    suspend fun fetchAppScript(app: App)
-}
-
 class GithubAppsFetcher(
     private val applicationFilesDir: String,
     private val connectionUtility: ConnectionUtility = ConnectionUtility(),
     private val logger: Logger = SentryLogger()
-) : RemoteAppsSource {
+) {
 
     // Allows destructing of the list of application elements
     operator fun <T> List<T>.component6() = get(5)
@@ -38,15 +28,15 @@ class GithubAppsFetcher(
     private var protocol = "https"
     private val hostname = "$protocol$baseUrl"
 
-    override fun getHostname(): String { return hostname }
+    fun getHostname(): String { return hostname }
 
     @Throws(IOException::class)
-    override suspend fun fetchAppsList(): List<App> {
+    suspend fun fetchAppsList(): List<App> = withContext(Dispatchers.IO) {
         val appsList = ArrayList<App>()
 
         val url = "$protocol$baseUrl/apps.txt"
         val numLinesToSkip = 1 // Skip first line defining schema
-        return try {
+        return@withContext try {
             val reader = BufferedReader(InputStreamReader(connectionUtility.getUrlInputStream(url)))
             reader.readLines().drop(numLinesToSkip).forEach {
                 val (name, category, filesystemRequired, supportsCli, supportsGui, isPaidApp, version) =
@@ -64,7 +54,7 @@ class GithubAppsFetcher(
         }
     }
 
-    override suspend fun fetchAppIcon(app: App) {
+    suspend fun fetchAppIcon(app: App) = withContext(Dispatchers.IO) {
         val directoryAndFilename = "${app.name}/${app.name}.png"
         val file = File("$applicationFilesDir/apps/$directoryAndFilename")
         file.parentFile!!.mkdirs()
@@ -78,7 +68,7 @@ class GithubAppsFetcher(
         outputStream.close()
     }
 
-    override suspend fun fetchAppDescription(app: App) {
+    suspend fun fetchAppDescription(app: App) = withContext(Dispatchers.IO) {
         val directoryAndFilename = "${app.name}/${app.name}.txt"
         val url = "$protocol$baseUrl/$directoryAndFilename"
         val file = File("$applicationFilesDir/apps/$directoryAndFilename")
@@ -88,7 +78,7 @@ class GithubAppsFetcher(
         file.writeText(contents)
     }
 
-    override suspend fun fetchAppScript(app: App) {
+    suspend fun fetchAppScript(app: App) = withContext(Dispatchers.IO) {
         val directoryAndFilename = "${app.name}/${app.name}.sh"
         val url = "$protocol$baseUrl/$directoryAndFilename"
         val file = File("$applicationFilesDir/apps/$directoryAndFilename")
@@ -96,5 +86,13 @@ class GithubAppsFetcher(
         file.createNewFile()
         val contents = URL(url).readText()
         file.writeText(contents)
+    }
+
+    private fun getBranchToDownloadAssetsFrom(assetType: String): String {
+        return when (assetType) {
+            "support" -> "staging"
+            "apps" -> "master"
+            else -> "master"
+        }
     }
 }
