+++ app/src/main/java/tech/ula/utils/DownloadUtility.kt
@@ -1,5 +1,8 @@
 package tech.ula.utils
 
+import android.app.DownloadManager
+import android.database.Cursor
+import android.net.Uri
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import org.rauschig.jarchivelib.Archiver
@@ -20,11 +23,10 @@ data class AssetDownloadFailure(val reason: DownloadFailureLocalizationData) : A
 class DownloadUtility(
     private val assetPreferences: AssetPreferences,
     private val downloadManagerWrapper: DownloadManagerWrapper,
-    private val applicationFilesDir: File,
-    scopedExternalRoot: File
+    private val ulaFiles: UlaFiles
 ) {
 
-    private val downloadDirectory = File(scopedExternalRoot, "downloads")
+    private val downloadDirectory = File(ulaFiles.scopedDir, "downloads")
 
     private val enqueuedDownloadIds = mutableSetOf<Long>()
     private val completedDownloadIds = mutableSetOf<Long>()
@@ -107,7 +109,7 @@ class DownloadUtility(
 
     @Throws(IOException::class)
     suspend fun prepareDownloadsForUse(archiverFactory: ArchiveFactoryWrapper = ArchiveFactoryWrapper()) = withContext(Dispatchers.IO) {
-        val stagingDirectory = File("${applicationFilesDir.path}/staging")
+        val stagingDirectory = File("${ulaFiles.filesDir.path}/staging")
         stagingDirectory.mkdirs()
         val downloadFiles = downloadDirectory.listFiles() ?: return@withContext
         downloadFiles.forEach {
@@ -122,7 +124,7 @@ class DownloadUtility(
 
     private suspend fun moveRootfsAssetInternal(rootFsFile: File) = withContext(Dispatchers.IO) {
         val (repo, filename, version) = rootFsFile.name.split("-", limit = 4)
-        val destinationDirectory = File("${applicationFilesDir.absolutePath}/$repo")
+        val destinationDirectory = File("${ulaFiles.filesDir.absolutePath}/$repo")
         val target = File("${destinationDirectory.absolutePath}/$filename")
 
         destinationDirectory.mkdirs()
@@ -143,7 +145,7 @@ class DownloadUtility(
     private suspend fun extractAssets(tarFile: File, stagingDirectory: File, archiverFactory: ArchiveFactoryWrapper) = withContext(Dispatchers.IO) {
         val (repo, filename, version) = tarFile.name.split("-", limit = 3)
         val stagingTarget = File("${stagingDirectory.absolutePath}/$filename")
-        val destination = File("${applicationFilesDir.path}/$repo")
+        val destination = File("${ulaFiles.filesDir.path}/$repo")
 
         tarFile.copyTo(stagingTarget, overwrite = true)
         tarFile.delete()
@@ -152,9 +154,79 @@ class DownloadUtility(
         archiver.extract(stagingTarget, destination)
         val extractedFiles = destination.listFiles() ?: return@withContext
         for (file in extractedFiles) {
-            makePermissionsUsable(destination.absolutePath, file.name)
+            ulaFiles.makePermissionsUsable(destination.absolutePath, file.name)
         }
         assetPreferences.setLatestDownloadVersion(repo, version)
+    }
+}
+
+class DownloadManagerWrapper(private val downloadManager: DownloadManager) {
+    fun generateDownloadRequest(url: String, destination: File): DownloadManager.Request {
+        val uri = Uri.parse(url)
+        val request = DownloadManager.Request(uri)
+        val destinationUri = Uri.fromFile(destination)
+        request.setAllowedNetworkTypes(DownloadManager.Request.NETWORK_WIFI or DownloadManager.Request.NETWORK_MOBILE)
+        request.setTitle(destination.name)
+        request.setDescription("Downloading ${destination.name.substringAfterLast("-")}.")
+        request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE)
+        request.setDestinationUri(destinationUri)
+        return request
+    }
+
+    fun enqueue(request: DownloadManager.Request): Long {
+        return downloadManager.enqueue(request)
+    }
+
+    private fun generateQuery(id: Long): DownloadManager.Query {
+        val query = DownloadManager.Query()
+        query.setFilterById(id)
+        return query
+    }
+
+    private fun generateCursor(query: DownloadManager.Query): Cursor {
+        return downloadManager.query(query)
+    }
+
+    fun downloadHasSucceeded(id: Long): Boolean {
+        val query = generateQuery(id)
+        val cursor = generateCursor(query)
+        if (cursor.moveToFirst()) {
+            val status = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_STATUS))
+            return status == DownloadManager.STATUS_SUCCESSFUL
+        }
+        return false
+    }
+
+    fun downloadHasFailed(id: Long): Boolean {
+        val query = generateQuery(id)
+        val cursor = generateCursor(query)
+        if (cursor.moveToFirst()) {
+            val status = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_STATUS))
+            return status == DownloadManager.STATUS_FAILED
+        }
+        return false
+    }
+
+    fun getDownloadFailureReason(id: Long): DownloadFailureLocalizationData {
+        val query = generateQuery(id)
+        val cursor = generateCursor(query)
+        if (cursor.moveToFirst()) {
+            val status = cursor.getInt(cursor.getColumnIndex(DownloadManager.COLUMN_REASON))
+            return DownloadFailureLocalizationData(resId = when (status) {
+                in 100..500 -> R.string.download_failure_http_error
+                1008 -> R.string.download_failure_cannot_resume
+                1007 -> R.string.download_failure_no_external_devices
+                1009 -> R.string.download_failure_destination_exists
+                1001 -> R.string.download_failure_unknown_file_error
+                1004 -> R.string.download_failure_http_processing
+                1006 -> R.string.download_failure_insufficient_external_storage
+                1005 -> R.string.download_failure_too_many_redirects
+                1002 -> R.string.download_failure_unhandled_http_response
+                1000 -> R.string.download_failure_unknown_error
+                else -> R.string.download_failure_missing_error
+            }, formatStrings = listOf("$status")) // Format strings only used for http_error
+        }
+        return DownloadFailureLocalizationData(R.string.download_failure_reason_not_found)
     }
 }
