+++ app/src/main/java/tech/ula/utils/UlaFiles.kt
@@ -22,6 +22,19 @@ class UlaFiles(
     val busybox = File(supportDir, "busybox")
     val proot = File(supportDir, "proot")
 
+    fun makePermissionsUsable(containingDirectoryPath: String, filename: String) {
+        val commandToRun = arrayListOf("chmod", "0777", filename)
+
+        val containingDirectory = File(containingDirectoryPath)
+        containingDirectory.mkdirs()
+
+        val pb = ProcessBuilder(commandToRun)
+        pb.directory(containingDirectory)
+
+        val process = pb.start()
+        process.waitFor()
+    }
+
     // Lib files must start with 'lib' and end with '.so.'
     private fun String.toSupportName(): String {
         return this.substringAfter("lib_").substringBeforeLast(".so")
